requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
geopandas==1.0.1
idna==3.10
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pyogrio==0.10.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
shapely==2.0.7
six==1.17.0
sodapy==2.2.0
tzdata==2025.1
urllib3==2.3.0
